requirements.txt
discord.py==2.3.0
pymongo==4.3.3
